Refactored code, issue with require causing tests to fail
