+++ apprentice_teacher.rb
@@ -1,3 +1,47 @@
+require_relative 'refactored_code'
+#Getting error: /Users/apprentice/Desktop/phase-1-checkpoint-challenge-3-teacher-refactoring/apprentice_teacher.rb:1:in
+#`require_relative': /Users/apprentice/Desktop/phase-1-checkpoint-challenge-3-teacher-refactoring/refactored_code.rb:38:
+#syntax error, unexpected keyword_end (SyntaxError)
+
+class ApprenticeTeacher < Person
+  attr_reader :salary, :target_raise
+
+  def initialize(options={})
+    super
+    @target_raise = 800
+  end
+
+  include Fivable
+  include Phasable
+  include Raisable
+  include Recievable
+
+  #this can be re-factored more, if we can alter the responses inside the actual class
+  def teach_stuff
+    response = ""
+    response += "Listen, class, this is how everything works. "
+    response += "*drops crazy knowledge bomb* "
+    response += "... You're welcome."
+    response
+  end
+
+  #this can be re-factored more, if we can alter the responses inside the actual class
+  def set_performance_rating(rating)
+    response = ""
+    if rating > 80
+      response = "Yay, I'm a great employee!"
+      receive_raise(@target_raise)
+    else
+      response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+    end
+    response
+  end
+
+  def attend_training_session
+    puts "Whoa. I know ruby-fu"
+  end
+end
+
 # class ApprenticeTeacher
 #   attr_reader :age, :salary, :phase, :target_raise
 #   attr_accessor :name
