+++ student.rb
@@ -1,3 +1,35 @@
+require_relative 'refactored_code'
+
+class Student < Person
+
+  def initialize(options = {})
+    super
+    @phase = 1
+  end
+
+  include Fivable
+
+  def set_phase(num)
+    response = ""
+    if num == @phase
+      response += "I'm doing phase #{@phase} again because "
+      response += "I put my learning first. I'm gonna rock it!"
+    else
+      response = "Oooh, phase #{num}. I hope I'm ready!"
+    end
+    @phase = num
+    response
+  end
+
+  def learn_stuff
+    response = ""
+    response += "WHOA! I've never thought of it quite like that before. "
+    response += "Now I feel like a genius!"
+    response
+  end
+end
+
+
 # class Student < Person
 #   attr_reader :age, :phase
 #   attr_accessor :name
