+++ teacher.rb
@@ -1,3 +1,41 @@
+require_relative 'refactored_code'
+
+class Teacher < Person
+  attr_reader :salary, :performance_rating, :target_raise
+
+  def initialize(options={})
+    super
+    @target_raise = 1000
+  end
+
+  include Fivable
+  include Phasable
+  include Raisable
+  include Recievable
+
+  #this can be re-factored more, if we can alter the responses inside the actual class
+  def teach_stuff
+    response = ""
+    response += "Listen, class, this is how everything works, fo SHO! "
+    response += "*drops flat-out insane knowledge bomb* "
+    response += "... You're welcome. *saunters away*"
+    response
+  end
+
+  #this can be re-factored more, if we can alter the responses inside the actual class
+  def set_performance_rating(rating)
+    response = ""
+    if rating > 90
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+end
+
 # class Teacher
 #   attr_reader :age, :salary, :phase, :performance_rating, :target_raise
 #   attr_accessor :name
